+++ shardingsphere-underlying/shardingsphere-route/src/test/resources/META-INF/services/org.apache.shardingsphere.underlying.route.hook.RoutingHook
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.sharding.route.fixture.RoutingHookFixture
+org.apache.shardingsphere.underlying.route.hook.fixture.RoutingHookFixture
